// src/types/ResumedOfficialServerInfo.ts
export default class ResumedOfficialServerInfo {
    id: string | undefined
    serverName: string | undefined;
    playersOnline: number| undefined;
    maxPlayers: number| undefined;
    mapName: string| undefined;
    day: string| undefined;
    ip: string| undefined;
    port: number| undefined;
    platforms: string[] | undefined
}